feat(user-list): implement panel
Implements new design's userlist panel taking the opportunity to
refactor the user dropdown menu logic and to break the user list item
component into smaller components.
