+++ bigbluebutton-html5/imports/ui/components/nav-bar/options-dropdown/container.jsx
@@ -16,7 +16,7 @@ import useCurrentUser from '/imports/ui/core/hooks/useCurrentUser';
 import useToggleVoice from '/imports/ui/components/audio/audio-graphql/hooks/useToggleVoice';
 import useWhoIsUnmuted from '/imports/ui/core/hooks/useWhoIsUnmuted';
 import Auth from '/imports/ui/services/auth';
-import { SET_AWAY } from '/imports/ui/components/user-list/user-list-content/user-participants/user-list-participants/user-actions/mutations';
+import { SET_AWAY } from '/imports/ui/components/user-list/user-list-participants/list-item/mutations';
 import {
   muteAway,
 } from '/imports/ui/components/audio/audio-graphql/audio-controls/input-stream-live-selector/service';
