+++ bigbluebutton-html5/imports/ui/components/profile-settings/component.tsx
@@ -18,7 +18,7 @@ import logger from '/imports/startup/client/logger';
 import MediaStreamUtils from '/imports/utils/media-stream-utils';
 import { getSettingsSingletonInstance } from '/imports/ui/services/settings';
 import useCurrentUser from '/imports/ui/core/hooks/useCurrentUser';
-import { SET_AWAY } from '/imports/ui/components/user-list/user-list-content/user-participants/user-list-participants/user-actions/mutations';
+import { SET_AWAY } from '../user-list/user-list-participants/list-item/mutations';
 import { muteAway } from '/imports/ui/components/audio/audio-graphql/audio-controls/input-stream-live-selector/service';
 import useToggleVoice from '/imports/ui/components/audio/audio-graphql/hooks/useToggleVoice';
 import useWhoIsUnmuted from '/imports/ui/core/hooks/useWhoIsUnmuted';
