+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-participants/intersection-watcher/intersectionWatcher.tsx
