+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-participants/list-item/skeleton/styles.ts
