+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-participants/list-item/user-item-toolbar/styles.ts
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+import { colorGrayIcons, colorGrayUserListToolbar } from '/imports/ui/stylesheets/styled-components/palette';
+
+const ToolbarContainer = styled.div`
+  border-radius: 1.5rem;
+  background-color: ${colorGrayUserListToolbar};
+  display: flex;
+  gap: 0.5rem;
+  padding: 0.25rem 1rem;
+  align-items: center;
+`;
+
+const ToolbarItem = styled.div`
+  cursor: pointer;
+  color: ${colorGrayIcons};
+`;
+
+const MoreItems = styled.div`
+  cursor: pointer;
+  color: ${colorGrayIcons};
+`;
+
+export default {
+  ToolbarContainer,
+  ToolbarItem,
+  MoreItems,
+};
