+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-participants/list-item/user-name-with-subs/types.ts
@@ -1,0 +1,11 @@
+import { IntlShape } from 'react-intl';
+import { User } from '/imports/ui/Types/user';
+import { LockSettings } from '/imports/ui/Types/meeting';
+import { UserListItemAdditionalInformationInterface } from 'bigbluebutton-html-plugin-sdk';
+
+export interface UserNameWithSubsProps {
+  intl: IntlShape;
+  subjectUser: User;
+  lockSettings: LockSettings;
+  userItemsFromPlugin: UserListItemAdditionalInformationInterface[]
+}
