+++ bigbluebutton-html5/public/locales/en.json
@@ -169,7 +169,8 @@
     "app.user.activityCheck": "User activity check",
     "app.user.activityCheck.label": "Check if user is still in session ({0})",
     "app.user.activityCheck.check": "Check",
-    "app.userList.usersTitle": "Users",
+    "app.userList.usersTitle": "User List",
+    "app.userList.close": "Close User List",
     "app.userList.profileTitle": "Profile",
     "app.userList.participantsTitle": "Participants",
     "app.userList.learningDashboardLabel": "Learning Analytics Dashboard",
@@ -193,6 +194,7 @@
     "app.userList.menuTitleContext": "Available options",
     "app.userList.chatListItem.unreadSingular": "One new message",
     "app.userList.chatListItem.unreadPlural": "{0} new messages",
+    "app.userList.more": "More",
     "app.userList.menu.away": "Set yourself as away",
     "app.userList.menu.notAway": "Set yourself as active",
     "app.userList.menu.chat.label": "Start a private chat",
